# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "D&D API",
    "version": "1.0.0",
    "description": "Documentación de la API para gestión de campañas y fichas de D&D"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Servidor Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Introduce el token JWT recibido en el login (sin el prefijo \"Bearer \")"
      }
    },
    "schemas": {
...
      "SystemApi": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID de MongoDB del sistema."
          },
          "name": {
            "type": "string",
            "description": "Nombre del sistema."
          },
          "description": {
            "type": "string",
            "description": "Descripción del sistema."
          },
          "publisher": {
            "type": "string",
            "description": "Nombre o ID del publicador."
          },
          "isOpen": {
            "type": "boolean",
            "description": "Indica si el sistema es abierto/público."
          },
          "isBase": {
            "type": "boolean",
            "description": "Indica si el sistema es una plantilla base (reglas únicamente)."
          },
          "parentId": {
            "type": "string",
            "description": "ID del sistema del que hereda (si lo tiene)."
          },
          "canEdit": {
            "type": "boolean",
            "description": "Indica si el usuario autenticado tiene permisos de edición."
          },
          "racesCount": {
            "type": "integer",
            "description": "Cantidad de razas asociadas (incluyendo heredadas)."
          },
          "languagesCount": {
            "type": "integer",
            "description": "Cantidad de idiomas asociados (incluyendo heredados)."
          },
          "traitsCount": {
            "type": "integer",
            "description": "Cantidad de rasgos asociados (incluyendo heredados)."
          },
          "globalModifierFormula": {
            "type": "string",
            "description": "Fórmula opcional para modificar características."
          },
          "initiativeBonusFormula": {
            "type": "string",
            "description": "Fórmula opcional para calcular el bono de iniciativa."
          },
          "defaultMinAttributeValue": {
            "type": "number"
          },
          "defaultMaxAttributeValue": {
            "type": "number"
          },
          "creationMinAttributeValue": {
            "type": "number"
          },
          "creationMaxAttributeValue": {
            "type": "number"
          },
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attribute"
            },
            "description": "Características vinculadas a este sistema (incluyendo heredadas)."
          }
        }
      },
      "TypeCrearSystem": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del sistema."
          },
          "description": {
            "type": "string",
            "description": "Descripción del sistema."
          },
          "isOpen": {
            "type": "boolean",
            "description": "Indica si es abierto."
          },
          "isBase": {
            "type": "boolean",
            "description": "Indica si es una plantilla base."
          },
          "parentId": {
            "type": "string",
            "description": "ID de MongoDB del sistema padre."
          },
          "globalModifierFormula": {
            "type": "string",
            "description": "Fórmula opcional para modificar características."
          },
          "initiativeBonusFormula": {
            "type": "string",
            "description": "Fórmula opcional para calcular el bono de iniciativa."
          },
          "defaultMinAttributeValue": {
            "type": "number"
          },
          "defaultMaxAttributeValue": {
            "type": "number"
          },
          "creationMinAttributeValue": {
            "type": "number"
          },
          "creationMaxAttributeValue": {
            "type": "number"
          }
        }
      },
      "TypeModificarSystem": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isOpen": {
            "type": "boolean"
          },
          "isBase": {
            "type": "boolean"
          },
          "parentId": {
            "type": "string"
          },
          "globalModifierFormula": {
            "type": "string",
            "description": "Fórmula opcional para modificar características."
          },
          "initiativeBonusFormula": {
            "type": "string",
            "description": "Fórmula opcional para calcular el bono de iniciativa."
          },
          "defaultMinAttributeValue": {
            "type": "number"
          },
          "defaultMaxAttributeValue": {
            "type": "number"
          },
          "creationMinAttributeValue": {
            "type": "number"
          },
          "creationMaxAttributeValue": {
            "type": "number"
          }
        }
      },
...
      "Attribute": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID de MongoDB de la característica."
          },
          "ruleset": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array de IDs o nombres de sistemas asociados."
          },
          "name": {
            "type": "string",
            "description": "Nombre de la característica (ej. Fuerza)."
          },
          "description": {
            "type": "string",
            "description": "Descripción detallada."
          },
          "key": {
            "type": "string",
            "description": "Clave identificadora única (ej. str, dex)."
          },
          "abbreviation": {
            "type": "string",
            "description": "Abreviatura (ej. FUE, DES)."
          },
          "icon": {
            "type": "string",
            "description": "Nombre o URL del icono opcional."
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/systems": {
      "get": {
        "summary": "Obtener los sistemas accesibles por el usuario",
        "tags": [
          "Sistemas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Listado de sistemas obtenidos exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SystemApi"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado."
          },
          "500": {
            "description": "Error del servidor."
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo sistema",
        "tags": [
          "Sistemas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TypeCrearSystem"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sistema creado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemApi"
                }
              }
            }
          },
          "400": {
            "description": "Nombre de sistema es obligatorio."
          },
          "401": {
            "description": "No autorizado."
          },
          "500": {
            "description": "Error del servidor."
          }
        }
      }
    },
    "/systems/{id}": {
      "put": {
        "summary": "Modificar un sistema existente",
        "tags": [
          "Sistemas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del sistema."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TypeModificarSystem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sistema modificado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemApi"
                }
              }
            }
          },
          "400": {
            "description": "Falta ID del sistema."
          },
          "401": {
            "description": "No autorizado."
          },
          "403": {
            "description": "No tienes permisos de edición o sistema no encontrado."
          },
          "500": {
            "description": "Error del servidor."
          }
        }
      },
      "delete": {
        "summary": "Realizar un borrado lógico de un sistema",
        "tags": [
          "Sistemas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del sistema a borrar."
          }
        ],
        "responses": {
          "204": {
            "description": "Sistema borrado exitosamente."
          },
          "400": {
            "description": "ID de sistema requerido."
          },
          "403": {
            "description": "No tienes permisos para borrar este sistema."
          },
          "404": {
            "description": "Sistema no encontrado."
          },
          "500": {
            "description": "Error del servidor."
          }
        }
      }
    },
    "/systems/{id}/restore": {
      "patch": {
        "summary": "Restaurar un sistema borrado lógicamente",
        "tags": [
          "Sistemas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del sistema a restaurar."
          }
        ],
        "responses": {
          "200": {
            "description": "Sistema restaurado exitosamente."
          },
          "400": {
            "description": "ID de sistema requerido."
          },
          "403": {
            "description": "No tienes permisos para restaurar este sistema."
          },
          "404": {
            "description": "Sistema no encontrado."
          },
          "500": {
            "description": "Error del servidor."
          }
        }
      }
    },
...
  },
...
}